requirements.txt
streamlit>=1.25.0
pdfplumber
python-docx
nltk
pandas
transformers
torch
sentencepiece
